+++ modules/4DCT/iADefectClassifier.cpp
@@ -1,7 +1,7 @@
 /*************************************  open_iA  ************************************ *
 * **********   A tool for visual analysis and processing of 3D CT images   ********** *
 * *********************************************************************************** *
-* Copyright (C) 2016-2019  C. Heinzl, M. Reiter, A. Reh, W. Li, M. Arikan, Ar. &  Al. *
+* Copyright (C) 2016-2020  C. Heinzl, M. Reiter, A. Reh, W. Li, M. Arikan, Ar. &  Al. *
 *                          Amirkhanov, J. Weissenböck, B. Fröhler, M. Schiwarth       *
 * *********************************************************************************** *
 * This program is free software: you can redistribute it and/or modify it under the   *
@@ -133,9 +133,9 @@ void iADefectClassifier::classify( FibersData* fibers, FeatureList* defects )
 			&& neighborFibersFF.size( ) >= 2 )
 		{
 			double minAngle = 2 * vtkMath::Pi(); // maximum possible angle
-			for( int i = 0; i < neighborFibersFF.size( ); ++i )
-			{				
-				for( int j = i + 1; j < neighborFibersFF.size( ); ++j )
+			for (size_t i = 0; i < neighborFibersFF.size( ); ++i)
+			{
+				for (size_t j = i + 1; j < neighborFibersFF.size( ); ++j)
 				{
 					double max[2], min[2];
 					max[0] = std::max( neighborFibersFF[i].startPoint[2], neighborFibersFF[i].endPoint[2] );
