+++ modules/CommonImageFilters/iACastImageFilter.cpp
@@ -1,7 +1,7 @@
 /*************************************  open_iA  ************************************ *
 * **********   A tool for visual analysis and processing of 3D CT images   ********** *
 * *********************************************************************************** *
-* Copyright (C) 2016-2019  C. Heinzl, M. Reiter, A. Reh, W. Li, M. Arikan, Ar. &  Al. *
+* Copyright (C) 2016-2020  C. Heinzl, M. Reiter, A. Reh, W. Li, M. Arikan, Ar. &  Al. *
 *                          Amirkhanov, J. Weissenböck, B. Fröhler, M. Schiwarth       *
 * *********************************************************************************** *
 * This program is free software: you can redistribute it and/or modify it under the   *
@@ -205,7 +205,7 @@ void convertToRGB(iAFilter * filter, QMap<QString, QVariant> const & params)
 	{
 		iAColorTheme const * theme = iAColorThemeManager::instance().theme(params["Color scheme"].toString());
 		labelToRGBFilter->ResetColors();
-		for (int c = 0; c < theme->size(); ++c)
+		for (size_t c = 0; c < theme->size(); ++c)
 		{
 			labelToRGBFilter->AddColor(theme->color(c).red(), theme->color(c).green(), theme->color(c).blue());
 		}
