+++ modules/GEMSe/iAParamHistogramData.h
@@ -1,7 +1,7 @@
 /*************************************  open_iA  ************************************ *
 * **********   A tool for visual analysis and processing of 3D CT images   ********** *
 * *********************************************************************************** *
-* Copyright (C) 2016-2019  C. Heinzl, M. Reiter, A. Reh, W. Li, M. Arikan, Ar. &  Al. *
+* Copyright (C) 2016-2020  C. Heinzl, M. Reiter, A. Reh, W. Li, M. Arikan, Ar. &  Al. *
 *                          Amirkhanov, J. Weissenböck, B. Fröhler, M. Schiwarth       *
 * *********************************************************************************** *
 * This program is free software: you can redistribute it and/or modify it under the   *
@@ -62,7 +62,7 @@ public:
 	double spacing() const override;
 	double const * xBounds() const override;
 	DataType const * yBounds() const override;
-	double binStart(int binNr) const override;
+	double binStart(size_t binNr) const override;
 	double mapValueToBin(double value) const;
 	double mapBinToValue(double bin) const;
 	iAValueType valueType() const override;
