+++ modules/Segmentation/iAGraphWeights.cpp
@@ -1,7 +1,7 @@
 /*************************************  open_iA  ************************************ *
 * **********   A tool for visual analysis and processing of 3D CT images   ********** *
 * *********************************************************************************** *
-* Copyright (C) 2016-2019  C. Heinzl, M. Reiter, A. Reh, W. Li, M. Arikan, Ar. &  Al. *
+* Copyright (C) 2016-2020  C. Heinzl, M. Reiter, A. Reh, W. Li, M. Arikan, Ar. &  Al. *
 *                          Amirkhanov, J. Weissenböck, B. Fröhler, M. Schiwarth       *
 * *********************************************************************************** *
 * This program is free software: you can redistribute it and/or modify it under the   *
@@ -44,7 +44,7 @@ iAEdgeWeightType iAGraphWeights::GetWeight(iAEdgeIndexType edgeIdx) const
 
 void iAGraphWeights::SetWeight(iAEdgeIndexType edgeIdx, iAEdgeWeightType weight)
 {
-	assert(edgeIdx < m_weights.size());
+	assert(edgeIdx < static_cast<unsigned int>(m_weights.size()));
 	m_weights[edgeIdx] = weight;
 }
